Main changed to call console. console changed to make more clear
