+++ src/Console.java
@@ -1,21 +1,8 @@
 import java.util.Scanner;
 
 public class Console {
-    private static Scanner in = new Scanner(System.in);
 
-    public static void welcome(){
-        //standaard welcome tekst
-        System.out.printf("Welcome by shipflex%n");
-        commandsStart();
-    }
-
-    public static void commandsStart(){
-        System.out.println("Commands: \'start\', \'onderhoud\', \'stop\'");
-        printInputText();
-    }
-    public static void printInputText(){
-        System.out.printf("tik een command in:%n");
-    }
+    p
     public static void commandsOnderhoud(){
         System.out.printf("Commands: \'optieLijst\', \'offerteMaken\', \'optieToevoegen\', \'OptieVerwijderen\',\'back\', \'stop\'%n");
         printInputText();
