+++ src/Main.java
@@ -1,5 +1,6 @@
 public class Main {
     public static void main(String[]args){
-        System.out.println("Hello World");
+        Console console = new Console();
+//        console.start();
     }
 }
